+++ app/src/main/java/com/lakeel/altla/vision/admanager/domain/usecase/appspace/DeleteMetaDataUseCase.java
@@ -1,5 +1,6 @@
 package com.lakeel.altla.vision.admanager.domain.usecase.appspace;
 
+import com.lakeel.altla.vision.admanager.ArgumentNullException;
 import com.lakeel.altla.vision.admanager.domain.repository.AppContentRepository;
 import com.lakeel.altla.vision.admanager.domain.repository.AppMetaDataRepository;
 
@@ -21,6 +22,8 @@ public final class DeleteMetaDataUseCase {
     }
 
     public Single<String> execute(String uuid) {
+        if (uuid == null) throw new ArgumentNullException("uuid");
+
         return appContentRepository.delete(uuid)
                                    .flatMap(appMetaDataRepository::delete)
                                    .subscribeOn(Schedulers.io());
