+++ app/src/main/java/com/lakeel/altla/vision/admanager/domain/usecase/tangospace/DeleteContentUseCase.java
@@ -1,5 +1,6 @@
 package com.lakeel.altla.vision.admanager.domain.usecase.tangospace;
 
+import com.lakeel.altla.vision.admanager.ArgumentNullException;
 import com.lakeel.altla.vision.admanager.domain.repository.TangoMetaDataRepository;
 
 import javax.inject.Inject;
@@ -17,6 +18,8 @@ public final class DeleteContentUseCase {
     }
 
     public Single<String> execute(String uuid) {
+        if (uuid == null) throw new ArgumentNullException("uuid");
+
         return tangoMetaDataRepository.delete(uuid)
                                       .subscribeOn(Schedulers.io());
     }
